Started refactor of system transaction handling.
