+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/components/state/StateManagementComponent.java
@@ -21,8 +21,6 @@ import com.swirlds.platform.components.common.output.NewSignedStateFromTransacti
 import com.swirlds.platform.components.common.output.RoundAppliedToStateConsumer;
 import com.swirlds.platform.components.common.output.SignedStateToLoadConsumer;
 import com.swirlds.platform.components.state.query.LatestSignedStateProvider;
-import com.swirlds.platform.components.transaction.system.PostConsensusSystemTransactionConsumer;
-import com.swirlds.platform.components.transaction.system.PreConsensusSystemTransactionConsumer;
 import com.swirlds.platform.state.signed.ReservedSignedState;
 import com.swirlds.platform.state.signed.SignedStateFinder;
 import com.swirlds.platform.state.signed.SignedStateInfo;
@@ -48,9 +46,7 @@ public interface StateManagementComponent
                 RoundAppliedToStateConsumer,
                 SignedStateToLoadConsumer,
                 NewSignedStateFromTransactionsConsumer,
-                LatestSignedStateProvider,
-                PreConsensusSystemTransactionConsumer,
-                PostConsensusSystemTransactionConsumer {
+                LatestSignedStateProvider {
 
     /**
      * Get a reserved instance of the latest immutable signed state. May be unhashed, may or may not have all required
