+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/components/transaction/system/PostConsensusSystemTransactionManager.java
@@ -16,19 +16,108 @@
 
 package com.swirlds.platform.components.transaction.system;
 
+import static com.swirlds.logging.LogMarker.EXCEPTION;
+
+import com.swirlds.common.system.NodeId;
+import com.swirlds.common.system.transaction.internal.SystemTransaction;
 import com.swirlds.platform.internal.ConsensusRound;
+import com.swirlds.platform.internal.EventImpl;
 import com.swirlds.platform.state.State;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 /**
- * Keeps track of system transaction post-consensus handling methods
+ * Routs system transactions to the appropriate handlers, post consensus.
  */
-@FunctionalInterface
-public interface PostConsensusSystemTransactionManager {
+public class PostConsensusSystemTransactionManager {
+
+    /**
+     * Class logger
+     */
+    private static final Logger logger = LogManager.getLogger(PostConsensusSystemTransactionManager.class);
+
+    /**
+     * The post-consensus handle methods that have been registered
+     */
+    private final Map<Class<?>, List<PostConsensusSystemTransactionHandler<SystemTransaction>>> handlers =
+            new HashMap<>();
+
+    /**
+     * Add a handle method
+     *
+     * @param clazz   the class of the transaction being handled
+     * @param handler a method to handle this transaction type
+     * @return this object
+     */
+    @SuppressWarnings("unchecked")
+    public <T extends SystemTransaction> PostConsensusSystemTransactionManager addHandler(
+            @NonNull final Class<T> clazz, @NonNull final PostConsensusSystemTransactionHandler<T> handler) {
+
+        handlers.computeIfAbsent(clazz, k -> new ArrayList<>())
+                .add((PostConsensusSystemTransactionHandler<SystemTransaction>) handler);
+
+        return this;
+    }
+
+    /**
+     * Pass an individual transaction to all handlers that want it
+     *
+     * @param state       the state
+     * @param creatorId   the id of the creator of the transaction
+     * @param transaction the transaction being handled
+     */
+    private void handleTransaction(
+            @NonNull final State state, @NonNull final NodeId creatorId, @NonNull final SystemTransaction transaction) {
+        Objects.requireNonNull(creatorId, "creatorId must not be null");
+        Objects.requireNonNull(transaction, "transaction must not be null");
+
+        final List<PostConsensusSystemTransactionHandler<SystemTransaction>> relevantHandlers =
+                handlers.get(transaction.getClass());
+
+        if (relevantHandlers == null) {
+            // no handlers exist that want this transaction type
+            return;
+        }
+
+        for (final PostConsensusSystemTransactionHandler<SystemTransaction> handler : relevantHandlers) {
+            try {
+                handler.handle(state, creatorId, transaction);
+            } catch (final RuntimeException e) {
+                logger.error(
+                        EXCEPTION.getMarker(),
+                        "Error while handling system transaction post consensus: "
+                                + "handler: {}, id: {}, transaction: {}, error: {}",
+                        handler,
+                        creatorId,
+                        transaction,
+                        e);
+            }
+        }
+    }
+
     /**
      * Handle a post-consensus round by passing each included system transaction to the registered handlers
      *
      * @param state a mutable state
      * @param round the post-consensus round
      */
-    void handleRound(State state, ConsensusRound round);
+    public void handleRound(@NonNull final State state, @NonNull final ConsensusRound round) {
+        // no post-consensus handling methods have been registered
+        if (handlers.isEmpty()) {
+            return;
+        }
+
+        Objects.requireNonNull(state);
+
+        for (final EventImpl event : round.getConsensusEvents()) {
+            event.systemTransactionIterator()
+                    .forEachRemaining(transaction -> handleTransaction(state, event.getCreatorId(), transaction));
+        }
+    }
 }
