+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/components/transaction/system/PreConsensusSystemTransactionManager.java
@@ -16,17 +16,101 @@
 
 package com.swirlds.platform.components.transaction.system;
 
+import static com.swirlds.logging.LogMarker.EXCEPTION;
+
+import com.swirlds.common.system.NodeId;
+import com.swirlds.common.system.transaction.internal.SystemTransaction;
 import com.swirlds.platform.internal.EventImpl;
+import edu.umd.cs.findbugs.annotations.NonNull;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 /**
- * Keeps track of system transaction pre-consensus handling methods
+ * Routs system transactions to the appropriate handlers, preconsensus.
  */
-@FunctionalInterface
-public interface PreConsensusSystemTransactionManager {
+public class PreConsensusSystemTransactionManager {
+
+    /**
+     * Class logger
+     */
+    private static final Logger logger = LogManager.getLogger(PreConsensusSystemTransactionManager.class);
+
+    /**
+     * The pre-consensus handle methods that have been registered
+     */
+    private final Map<Class<?>, List<PreConsensusSystemTransactionHandler<SystemTransaction>>> handlers =
+            new HashMap<>();
+
+    /**
+     * Add a handle method
+     *
+     * @param clazz   the class of the transaction being handled
+     * @param handler a method to handle this transaction type
+     * @return this object
+     */
+    @SuppressWarnings("unchecked")
+    public <T extends SystemTransaction> PreConsensusSystemTransactionManager addHandler(
+            @NonNull final Class<T> clazz, @NonNull final PreConsensusSystemTransactionHandler<T> handler) {
+
+        Objects.requireNonNull(clazz);
+        Objects.requireNonNull(handler);
+
+        handlers.computeIfAbsent(clazz, k -> new ArrayList<>())
+                .add((PreConsensusSystemTransactionHandler<SystemTransaction>) handler);
+
+        return this;
+    }
+
+    /**
+     * Pass an individual transaction to all handlers that want it
+     *
+     * @param creatorId   the id of the creator of the transaction
+     * @param transaction the transaction being handled
+     */
+    private void handleTransaction(@NonNull final NodeId creatorId, @NonNull final SystemTransaction transaction) {
+        Objects.requireNonNull(creatorId, "creatorId must not be null");
+        Objects.requireNonNull(transaction, "transaction must not be null");
+
+        final List<PreConsensusSystemTransactionHandler<SystemTransaction>> relevantHandlers =
+                handlers.get(transaction.getClass());
+
+        if (relevantHandlers == null) {
+            // no handlers exist that want this transaction type in this stage
+            return;
+        }
+
+        for (final PreConsensusSystemTransactionHandler<SystemTransaction> handler : relevantHandlers) {
+            try {
+                handler.handle(creatorId, transaction);
+            } catch (final RuntimeException e) {
+                logger.error(
+                        EXCEPTION.getMarker(),
+                        "Error while handling system transaction pre-consensus: handler: {}, id: {}, transaction: {}, error: {}",
+                        handler,
+                        creatorId,
+                        transaction,
+                        e);
+            }
+        }
+    }
+
     /**
      * Handle a pre-consensus event by passing each included system transaction to the registered handlers
      *
      * @param event the pre-consensus event
      */
-    void handleEvent(EventImpl event);
+    public void handleEvent(@NonNull final EventImpl event) {
+        // no pre-consensus handling methods have been registered
+        if (handlers.isEmpty()) {
+            return;
+        }
+
+        event.systemTransactionIterator()
+                .forEachRemaining(transaction -> handleTransaction(event.getCreatorId(), transaction));
+    }
 }
