+++ platform-sdk/swirlds-platform-core/src/main/java/module-info.java
@@ -114,7 +114,6 @@ module com.swirlds.platform {
             info.picocli;
 
     exports com.swirlds.platform.components.transaction;
-    exports com.swirlds.platform.components.transaction.system.internal;
     exports com.swirlds.platform.components.transaction.system;
     exports com.swirlds.platform.event.preconsensus;
     exports com.swirlds.platform.gossip.sync.protocol;
