+++ platform-sdk/swirlds-unit-tests/core/swirlds-platform-test/src/test/java/com/swirlds/platform/test/eventflow/SystemTransactionTracker.java
@@ -20,21 +20,15 @@ import com.swirlds.common.system.NodeId;
 import com.swirlds.common.system.transaction.Transaction;
 import com.swirlds.common.system.transaction.internal.SystemTransactionPing;
 import com.swirlds.common.test.fixtures.TransactionUtils;
-import com.swirlds.platform.components.transaction.system.PostConsensusSystemTransactionConsumer;
-import com.swirlds.platform.components.transaction.system.PostConsensusSystemTransactionTypedHandler;
-import com.swirlds.platform.components.transaction.system.PreConsensusSystemTransactionConsumer;
-import com.swirlds.platform.components.transaction.system.PreConsensusSystemTransactionTypedHandler;
 import com.swirlds.platform.state.State;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.HashMap;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Set;
 
-public class SystemTransactionTracker
-        implements PreConsensusSystemTransactionConsumer, PostConsensusSystemTransactionConsumer, Failable {
+public class SystemTransactionTracker implements Failable {
 
     private final Map<NodeId, Integer> preConsByCreator = new HashMap<>();
     private final Set<Transaction> preConsensusTransactions = new HashSet<>();
@@ -115,15 +109,16 @@ public class SystemTransactionTracker
         return failureMsg.toString();
     }
 
-    @Override
-    public List<PreConsensusSystemTransactionTypedHandler<?>> getPreConsensusHandleMethods() {
-        return List.of(new PreConsensusSystemTransactionTypedHandler<>(
-                SystemTransactionPing.class, this::handlePreConsensusSystemTransaction));
-    }
-
-    @Override
-    public List<PostConsensusSystemTransactionTypedHandler<?>> getPostConsensusHandleMethods() {
-        return List.of(new PostConsensusSystemTransactionTypedHandler<>(
-                SystemTransactionPing.class, this::handlePostConsensusSystemTransaction));
-    }
+    // TODO
+    //    @Override
+    //    public List<PreConsensusSystemTransactionTypedHandler<?>> getPreConsensusHandleMethods() {
+    //        return List.of(new PreConsensusSystemTransactionTypedHandler<>(
+    //                SystemTransactionPing.class, this::handlePreConsensusSystemTransaction));
+    //    }
+    //
+    //    @Override
+    //    public List<PostConsensusSystemTransactionTypedHandler<?>> getPostConsensusHandleMethods() {
+    //        return List.of(new PostConsensusSystemTransactionTypedHandler<>(
+    //                SystemTransactionPing.class, this::handlePostConsensusSystemTransaction));
+    //    }
 }
